feat: resolve opt in nft to control
