+++ src/utils/server/thirdweb/onLogin.ts
@@ -871,11 +871,7 @@ async function triggerPostLoginUserActionSteps({
     })
     log(`triggerPostLoginUserActionSteps: opt in user action created`)
 
-    const signUpFlowExperimentVariant =
-      localUser?.persisted?.experiments?.gh02_SWCSignUpFlowExperiment
-    if (signUpFlowExperimentVariant === 'control') {
-      await claimNFTAndSendEmailNotification(optInUserAction, userCryptoAddress)
-    }
+    await claimNFTAndSendEmailNotification(optInUserAction, userCryptoAddress)
 
     if (embeddedWalletUserDetails?.phone) {
       await smsActions.optInUser(embeddedWalletUserDetails.phone, user)
